fix: MatchMedia types
