--- packages/react-ui-kit/src/Layout/MatchMedia.jsx
@@ -1,50 +1,0 @@
-import {QUERY} from '../mediaQueries';
-import React from 'react';
-import {defaultProps} from 'recompose';
-
-interface MatchMediaProps {
-  not?: boolean;
-  query: string;
-}
-
-class MatchMedia extends React.PureComponent<MatchMediaProps & React.HTMLAttributes<HTMLElement>> {
-  static defaultProps = {
-    not: false,
-  };
-  matchMedia: MediaQueryList;
-  state: {
-    isMatching: boolean,
-  };
-  constructor(props) {
-    super(props);
-    this.matchMedia = window.matchMedia(`(${props.query})`);
-    this.state = {
-      isMatching: this.matchMedia.matches,
-    };
-  }
-
-  updateState = () => this.setState({isMatching: this.matchMedia.matches});
-
-  componentDidMount() {
-    this.matchMedia.addListener(this.updateState);
-  }
-
-  componentWillUnmount() {
-    this.matchMedia.removeListener(this.updateState);
-  }
-
-  render() {
-    const isMatching = this.props.not ? !this.state.isMatching : this.state.isMatching;
-    return isMatching ? this.props.children : null;
-  }
-}
-
-const IsDesktop = defaultProps({query: QUERY.desktop})(MatchMedia);
-const IsDesktopXL = defaultProps({query: QUERY.desktopXL})(MatchMedia);
-const IsMobile = defaultProps({query: QUERY.mobile})(MatchMedia);
-const IsMobileUp = defaultProps({query: QUERY.mobileUp})(MatchMedia);
-const IsTablet = defaultProps({query: QUERY.tablet})(MatchMedia);
-const IsTabletDown = defaultProps({query: QUERY.tabletDown})(MatchMedia);
-const IsTabletUp = defaultProps({query: QUERY.tabletUp})(MatchMedia);
-
-export {MatchMedia, IsDesktop, IsDesktopXL, IsMobile, IsMobileUp, IsTablet, IsTabletDown, IsTabletUp};
